Split InjectionPointTest into two test classes.
