+++ impl/src/main/java/org/jboss/cdi/tck/tests/full/lookup/injectionpoint/AnimalDecorator1.java
@@ -9,32 +9,30 @@
  * You may obtain a copy of the License at
  * http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,  
+ * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.cdi.tck.tests.lookup.injectionpoint;
+package org.jboss.cdi.tck.tests.full.lookup.injectionpoint;
 
-import jakarta.enterprise.context.Dependent;
+import jakarta.decorator.Decorator;
+import jakarta.decorator.Delegate;
 import jakarta.inject.Inject;
 
-/**
- * Test bean to inject a bean using injection point metadata into a constructor
- * 
- * @author David Allen
- * 
- */
-@Dependent
-public class ConstructorInjectionPointBean {
-    private BeanWithInjectionPointMetadata injectedBean;
+@Decorator
+public abstract class AnimalDecorator1 implements Animal {
 
     @Inject
-    public ConstructorInjectionPointBean(BeanWithInjectionPointMetadata injectedBean) {
-        this.injectedBean = injectedBean;
+    @Delegate
+    private Animal animal;
+
+    @Inject
+    @SuppressWarnings("unused")
+    private Toy toy;
+
+    public String hello() {
+        return animal.hello() + "!";
     }
 
-    public BeanWithInjectionPointMetadata getInjectedBean() {
-        return injectedBean;
-    }
 }
