+++ impl/src/main/java/org/jboss/cdi/tck/tests/full/lookup/injectionpoint/AnimalStereotype.java
@@ -9,30 +9,28 @@
  * You may obtain a copy of the License at
  * http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
+ * distributed under the License is distributed on an "AS IS" BASIS,  
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.cdi.tck.tests.lookup.injectionpoint;
+package org.jboss.cdi.tck.tests.full.lookup.injectionpoint;
 
-import jakarta.decorator.Decorator;
-import jakarta.decorator.Delegate;
-import jakarta.inject.Inject;
+import static java.lang.annotation.ElementType.FIELD;
+import static java.lang.annotation.ElementType.METHOD;
+import static java.lang.annotation.ElementType.TYPE;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
-@Decorator
-public abstract class AnimalDecorator1 implements Animal {
+import jakarta.enterprise.context.RequestScoped;
+import jakarta.enterprise.inject.Stereotype;
 
-    @Inject
-    @Delegate
-    private Animal animal;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
 
-    @Inject
-    @SuppressWarnings("unused")
-    private Toy toy;
-
-    public String hello() {
-        return animal.hello() + "!";
-    }
+@Stereotype
+@Target({ TYPE, METHOD, FIELD })
+@Retention(RUNTIME)
+@RequestScoped
+public @interface AnimalStereotype {
 
 }
