+++ impl/src/main/java/org/jboss/cdi/tck/tests/full/lookup/injectionpoint/Cat.java
@@ -14,12 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.cdi.tck.tests.lookup.injectionpoint;
+package org.jboss.cdi.tck.tests.full.lookup.injectionpoint;
 
+import jakarta.enterprise.context.Dependent;
 import jakarta.enterprise.inject.spi.BeanManager;
 import jakarta.enterprise.inject.spi.InjectionPoint;
 import jakarta.inject.Inject;
 
+@Dependent
 public class Cat implements Animal {
     @Inject
     private InjectionPoint injectionPoint;
