+++ impl/src/main/java/org/jboss/cdi/tck/tests/full/lookup/injectionpoint/Toy.java
@@ -9,33 +9,22 @@
  * You may obtain a copy of the License at
  * http://www.apache.org/licenses/LICENSE-2.0
  * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
+ * distributed under the License is distributed on an "AS IS" BASIS,  
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.jboss.cdi.tck.tests.lookup.injectionpoint;
+package org.jboss.cdi.tck.tests.full.lookup.injectionpoint;
 
-import jakarta.decorator.Decorator;
-import jakarta.decorator.Delegate;
+import jakarta.enterprise.inject.spi.InjectionPoint;
 import jakarta.inject.Inject;
 
-@Decorator
-public abstract class AnimalDecorator2 implements Animal {
+public class Toy {
 
     @Inject
-    @Delegate
-    private Animal animal;
+    private InjectionPoint injectionPoint;
 
-    @Inject
-    private Toy toy;
-
-    public String hello() {
-        return animal.hello() + "!";
-    }
-
-    @Override
-    public Toy getToy() {
-        return toy;
+    public InjectionPoint getInjectionPoint() {
+        return injectionPoint;
     }
 }
