+++ impl/src/main/java/org/jboss/cdi/tck/tests/lookup/injectionpoint/BazProducer.java
@@ -22,6 +22,7 @@ import jakarta.enterprise.context.Dependent;
 import jakarta.enterprise.inject.Produces;
 import jakarta.enterprise.inject.spi.InjectionPoint;
 
+@Dependent
 public class BazProducer {
 
     @Produces
