+++ impl/src/main/java/org/jboss/cdi/tck/tests/lookup/injectionpoint/InjectionPointTest.java
@@ -18,7 +18,6 @@ package org.jboss.cdi.tck.tests.lookup.injectionpoint;
 
 import static org.jboss.cdi.tck.cdi.Sections.CONTEXTUAL_REFERENCE;
 import static org.jboss.cdi.tck.cdi.Sections.INJECTION_POINT;
-import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
 
 import jakarta.enterprise.context.Dependent;
@@ -26,13 +25,11 @@ import jakarta.enterprise.inject.Default;
 import jakarta.enterprise.inject.spi.AnnotatedField;
 import jakarta.enterprise.inject.spi.AnnotatedParameter;
 import jakarta.enterprise.inject.spi.Bean;
-import jakarta.enterprise.inject.spi.Decorator;
 import jakarta.enterprise.inject.spi.InjectionPoint;
 import org.jboss.arquillian.container.test.api.Deployment;
 import org.jboss.cdi.tck.AbstractTest;
 import org.jboss.cdi.tck.shrinkwrap.WebArchiveBuilder;
 import org.jboss.shrinkwrap.api.spec.WebArchive;
-import org.jboss.shrinkwrap.impl.BeansXml;
 import org.jboss.test.audit.annotations.SpecAssertion;
 import org.jboss.test.audit.annotations.SpecAssertions;
 import org.jboss.test.audit.annotations.SpecVersion;
@@ -42,9 +39,6 @@ import java.lang.annotation.Annotation;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
-import java.lang.reflect.Type;
-import java.util.Collections;
-import java.util.List;
 import java.util.Set;
 
 /**
@@ -60,7 +54,6 @@ public class InjectionPointTest extends AbstractTest {
     public static WebArchive createTestArchive() {
         return new WebArchiveBuilder()
                 .withTestClassPackage(InjectionPointTest.class)
-                .withBeansXml(new BeansXml().decorators(AnimalDecorator1.class, AnimalDecorator2.class, AnimalDecorator3.class))
                 .build();
     }
 
@@ -181,22 +174,6 @@ public class InjectionPointTest extends AbstractTest {
     }
 
     @Test
-    @SpecAssertion(section = INJECTION_POINT, id = "eb")
-    public void testPassivationCapability() throws Exception {
-        InjectionPoint ip1 = getContextualReference(FieldInjectionPointBean.class).getInjectedBean().getInjectedMetadata();
-        InjectionPoint ip2 = getContextualReference(MethodInjectionPointBean.class).getInjectedBean().getInjectedMetadata();
-        InjectionPoint ip3 = getContextualReference(ConstructorInjectionPointBean.class).getInjectedBean().getInjectedMetadata();
-
-        ip1 = (InjectionPoint) activate(passivate(ip1));
-        ip2 = (InjectionPoint) activate(passivate(ip2));
-        ip3 = (InjectionPoint) activate(passivate(ip3));
-
-        assert ip1.getType().equals(BeanWithInjectionPointMetadata.class);
-        assert ip2.getType().equals(BeanWithInjectionPointMetadata.class);
-        assert ip3.getType().equals(BeanWithInjectionPointMetadata.class);
-    }
-
-    @Test
     @SpecAssertion(section = INJECTION_POINT, id = "ea")
     public void testApiTypeInjectionPoint() {
         // Get an instance of the bean which has another bean injected into it
@@ -226,47 +203,4 @@ public class InjectionPointTest extends AbstractTest {
         assert !ip1.isTransient();
         assert ip2.isTransient();
     }
-
-    /**
-     * CDI-78 reopened.
-     *
-     * Base of this test was originally part of CDITCK-138 but disappeared during branch merge.
-     */
-    @Test
-    @SpecAssertion(section = INJECTION_POINT, id = "dba")
-    public void testIsDelegate() {
-
-        assert !getContextualReference(FieldInjectionPointBean.class).getInjectedBean().getInjectedMetadata().isDelegate();
-
-        Cat cat = getContextualReference(Cattery.class).getCat();
-        // Cat is decorated
-        assert cat.hello().equals("hello!!!");
-        assert cat.getBeanManager() != null;
-        assert cat.getInjectionPoint() != null;
-        assert !cat.getInjectionPoint().isDelegate();
-
-        List<Decorator<?>> animalDecorators = getCurrentManager().resolveDecorators(Collections.<Type> singleton(Animal.class));
-        assert animalDecorators.size() == 3;
-        for (Decorator<?> animalDecorator : animalDecorators) {
-            // Decorator has two injection points - metadata and delegate
-            assert animalDecorator.getInjectionPoints().size() == 2;
-
-            for (InjectionPoint injectionPoint : animalDecorator.getInjectionPoints()) {
-                if (injectionPoint.getType().equals(InjectionPoint.class)) {
-                    assertFalse(injectionPoint.isDelegate());
-                } else if (injectionPoint.getType().equals(Animal.class)) {
-                    assertTrue(injectionPoint.isDelegate());
-                } else if (injectionPoint.getType().equals(Toy.class)) {
-                    assertFalse(injectionPoint.isDelegate());
-                } else {
-                    // Unexpected injection point type
-                    assert false;
-                }
-            }
-        }
-        Toy toy = cat.getToy();
-        assert toy.getInjectionPoint() != null;
-        assert toy.getInjectionPoint().getBean().getBeanClass().equals(AnimalDecorator2.class);
-    }
-
 }
