+++ impl/src/main/java/org/jboss/cdi/tck/tests/lookup/injectionpoint/MethodInjectionPointBean.java
@@ -17,6 +17,7 @@
 
 package org.jboss.cdi.tck.tests.lookup.injectionpoint;
 
+import jakarta.enterprise.context.Dependent;
 import jakarta.inject.Inject;
 
 /**
@@ -25,6 +26,7 @@ import jakarta.inject.Inject;
  * @author David Allen
  * 
  */
+@Dependent
 public class MethodInjectionPointBean {
     private BeanWithInjectionPointMetadata injectedBean;
